Move functypes, might rename files again
